pyproject.toml
[project]
name = "church-manager-v4"
version = "0.1.0"
description = "Church Manager v4 API"
authors = [
    {name = "Your Name",email = "your.email@example.com"}
]
readme = "README.md"
requires-python = "^3.13"
dependencies = [
    "fastapi (>=0.116.1,<0.117.0)",
    "uvicorn (>=0.35.0,<0.36.0)",
    "sqlalchemy (>=2.0.36,<3.0.0)",
    "asyncpg (>=0.30.0,<0.31.0)",
    "alembic (>=1.15.0,<2.0.0)",
    "pydantic (>=2.10.3,<3.0.0)",
    "pydantic-settings (>=2.7.0,<3.0.0)",
    "python-dotenv (>=1.0.1,<2.0.0)",
    "greenlet (>=3.2.4,<4.0.0)",
    "httpx (>=0.27.0,<0.28.0)",
    "pytest (>=8.3.0,<9.0.0)",
    "pytest-asyncio (>=0.25.0,<0.26.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
